+++ adapter/amazon/src/main/java/org/bidon/amazon/impl/AmazonRewardedImpl.kt
@@ -97,7 +97,7 @@ internal class AmazonRewardedImpl(
                             AdEvent.PaidRevenue(
                                 ad = it,
                                 adValue = AdValue(
-                                    adRevenue = adParams.price,
+                                    adRevenue = adParams.price / 1000.0,
                                     currency = AdValue.USD,
                                     Precision.Precise
                                 )